requirements.txt
aiogram==2.25.2
aiohttp
aiosqlite
fastapi
uvicorn
python-dotenv
